// src/components/DataGrid/types/PredefinedFormat.ts
/** DevExtreme's predefined format names. */
export type PredefinedFormat =
  | 'currency'
  | 'fixedPoint'
  | 'decimal'
  | 'percent'
  | 'exponential'
  | 'largeNumber'
  | 'thousands'
  | 'millions'
  | 'billions'
  | 'trillions'
  | 'shortDate'
  | 'longDate'
  | 'shortTime'
  | 'longTime'
  | 'shortDateShortTime'
  | 'longDateLongTime'
  | 'monthAndDay'
  | 'monthAndYear'
  | 'quarterAndYear'
  | 'year'
  | 'quarter'
  | 'month'
  | 'day'
  | 'dayOfWeek'
  | 'hour'
  | 'minute'
  | 'second'
  | 'millisecond'